Add functionality on what logged in user can edaccess

+++ booking/views.py
@@ -8,7 +8,6 @@ from django.views import View, generic
 from django.contrib import messages
 from django.views.generic.edit import FormView,  UpdateView
 from django.views.generic import ListView, DetailView
-from django.views.generic.base import TemplateView  #Builds view classes that render templates.
 from .models import Reservation, Cancel
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
@@ -16,6 +15,8 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic.edit import CreateView
 from django.contrib.auth import authenticate, login, logout
 from django.urls import reverse_lazy
+from datetime import date, datetime
+from django.http import HttpResponseForbidden
 
 
 
@@ -28,18 +29,27 @@ def starting_page(request):
 
 class ReserveView(LoginRequiredMixin, CreateView):
     """
-   This class will save the data entered by the user after submission and redirect to success url!
+    This class will save the data entered by the user after submission and redirect to success url!
     """
     model = Reservation
     form_class = ReserveForm
     template_name = 'booking/reserve.html'
-   
-   
 
     def form_valid(self, form):
+        # Extract the cleaned data
+        booked_time = form.cleaned_data['time_of_day']
+        booking_date = form.cleaned_data['date_of_month']
+        current_time = datetime.now().time()
+
+        # Check if the booking date is today and the time is in the past
+        if booking_date == date.today() and booked_time < current_time:
+            form.add_error('time_of_day', 'The time cannot be in the past for today\'s bookings.')
+            return self.form_invalid(form)
+
+        # Associate the form instance with the user
         form.instance.user = self.request.user
         messages.success(self.request, 'Thank you for making your reservation with us, a table has been reserved for you.')
-       
+        
         return super().form_valid(form)
 
     def get_success_url(self):
@@ -58,8 +68,7 @@ class ReserveViewList(LoginRequiredMixin,  generic.ListView):
     def get_queryset(self):
         return Reservation.objects.filter(user=self.request.user)
 
-        # user = self.request.user
-        # return Reservation.objects.filter(user=user), to get only one reservation by a user
+
 
 
 def single_reservation(request, pk):
@@ -68,9 +77,11 @@ def single_reservation(request, pk):
     """
    
     reservation = get_object_or_404(Reservation, pk=pk)
-    # print("reservation = ", order.name_user)
 
-
+    if reservation.user != request.user:
+        return HttpResponseForbidden("You are not allowed to access this reservation.")
+        
+    
     return render(request,'booking/single_reservation.html', {'reservation': reservation})
  
 
@@ -79,21 +90,30 @@ def single_reservation(request, pk):
 
 @login_required
 def cancel_reservation(request, reservation_id ):
-    """ This view will cancel a booking. """
+    """ This view will cancel a booking. 
+    """
+
     reservation = get_object_or_404(Reservation, pk=reservation_id)
-    
+
+    if reservation.user != request.user:
+        return HttpResponseForbiden("You are not allowed to delete this reservation.")
+
     if reservation.user == request.user:
         reservation.delete()
         messages.add_message(request, messages.SUCCESS, 'Reservation has been deleted successfully!')
     else:
         messages.add_message(request, messages.ERROR, 'There was an error cancelling your reservation,please check the details entered and try again. If the problem persists please contact the restaurant via telephone. ')
-    return HttpResponseRedirect(reverse('booking:starting_page'))
+    return HttpResponseRedirect(reverse('booking:reserve'))
 
 
    
 def update_reservation(request, reservation_id):
     change = get_object_or_404(Reservation, pk=reservation_id)
     form = EditForm(instance=change)
+
+    if change.user != request.user:
+        return HttpResponseForbiden("You are not allowed to edit this reservation.")
+
     if request.method == 'POST':
         form = ReserveForm(request.POST, instance= change)
         if form.is_valid() and change.user == request.user:
